Add tpt-t-link transport layer and AI input source
- Add tpt-t-link reliable transport: io_uring service loop, stream
  multiplexer (mux), mesh peer routing, reliability/retransmission,
  backpressure, and CRC framing.
- Add AI-source input backend (ai_source) feeding commands into the
  lock-free pipeline alongside real-device sources.
- Extend tpt-t-core command serialization with new command variants.
- Harden input subsystem: evdev parsing, HID mapping, haptics routing,
  and pipeline batching; broaden e2e input coverage.
- Strengthen safety loop with measured-state guards and recovery paths.
- Refresh todo/status notes.

+++ crates/tpt-t-input/benches/poll_bench.rs
@@ -24,7 +24,7 @@ fn report() -> ControllerReport {
 fn ingest(map: &ControllerMap, input: &SpscRing<ControlCommand>, i: u64) {
     let mut cmd = ControlCommand::zeroed(Mode::FullTeleop);
     map.apply(&report(), &mut cmd);
-    cmd.seq = i as u32;
+    cmd.seq = i;
     let _ = input.push(cmd);
 }
